Refactor Filter
