+++ src/pages/Main.js
@@ -1,10 +1,10 @@
 /* eslint-disable react/prop-types */
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Dropdown } from 'react-bootstrap';
 import Navbar from '../components/Navbar';
 import Total from '../components/Total';
 import Country from '../components/Country';
+import Filter from '../components/Filter';
 import { getContries, filterCountries } from '../redux/countries/countries';
 import '../css/Main.css';
 
@@ -27,22 +27,13 @@ const Main = () => {
     dispatch(filterCountries(e));
   };
 
-  const letters = 'ABCDEFGHIJKLMNOPQRSTUVWYZ';
-  const buttonList = letters.split('').map((letter) => <Dropdown.Item className="text-center" key={letter} eventKey={letter}>{letter}</Dropdown.Item>);
   return (
     <>
       <Navbar title="World stats" />
       <section className="px-1">
         <Total name="World" total={todayConfirmed} filter={handleClick} />
         <div>
-          <Dropdown onSelect={handleClick}>
-            <Dropdown.Toggle className="filter-button border-0 w-100" id="dropdown-basic">
-              Filter
-            </Dropdown.Toggle>
-            <Dropdown.Menu className="w-100">
-              {buttonList}
-            </Dropdown.Menu>
-          </Dropdown>
+          <Filter handleSelect={handleClick} />
           <ul className="d-flex flex-row flex-wrap w-100 px-0">
             {list}
           </ul>
